requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
whitenoise==6.6.0
gunicorn==21.2.0
django-redis==5.4.0
redis==5.0.3
